Test User, Repozytoria i drobne zmiany w encjach

+++ src/main/java/com/kodilla/ecommercee/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package com.kodilla.ecommercee.repository;
 import com.kodilla.ecommercee.domain.Order;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends CrudRepository<Order, Long> {
+public interface OrderRepository extends CrudRepository<Order, Integer> {
 }
